+++ sme/src/sme_common.cpp
@@ -2,31 +2,31 @@
 
 namespace sme {
 
-PyImageRgb toPyImageRgb(const QImage &img) {
-  auto h = static_cast<std::size_t>(img.height());
-  auto w = static_cast<std::size_t>(img.width());
-  PyImageRgb v(h, std::vector<std::vector<int>>(w, std::vector<int>(3, 0)));
-  for (std::size_t y = 0; y < h; ++y) {
-    for (std::size_t x = 0; x < w; ++x) {
-      auto c = img.pixel(static_cast<int>(x), static_cast<int>(y));
-      v[y][x][0] = qRed(c);
-      v[y][x][1] = qGreen(c);
-      v[y][x][2] = qBlue(c);
+// 3d array with triplet of r,g,b uint8 values for each pixel
+pybind11::array toPyImageRgb(const QImage &img) {
+  pybind11::array_t<std::uint8_t> a({img.height(), img.width(), 3});
+  auto r{a.mutable_unchecked<3>()};
+  for (int y = 0; y < img.height(); ++y) {
+    for (int x = 0; x < img.width(); ++x) {
+      auto c{img.pixel(x, y)};
+      r(y, x, 0) = static_cast<std::uint8_t>(qRed(c));
+      r(y, x, 1) = static_cast<std::uint8_t>(qGreen(c));
+      r(y, x, 2) = static_cast<std::uint8_t>(qBlue(c));
     }
   }
-  return v;
+  return a;
 }
 
-PyImageMask toPyImageMask(const QImage &img) {
-  auto h = static_cast<std::size_t>(img.height());
-  auto w = static_cast<std::size_t>(img.width());
-  PyImageMask v(h, std::vector<bool>(w, false));
-  for (std::size_t y = 0; y < h; ++y) {
-    for (std::size_t x = 0; x < w; ++x) {
-      v[y][x] = img.pixelIndex(static_cast<int>(x), static_cast<int>(y)) > 0;
+// 2d array of bool for each pixel
+pybind11::array toPyImageMask(const QImage &img) {
+  pybind11::array_t<bool> a({img.height(), img.width()});
+  auto r{a.mutable_unchecked<2>()};
+  for (int y = 0; y < img.height(); ++y) {
+    for (int x = 0; x < img.width(); ++x) {
+      r(y, x) = img.pixelIndex(x, y) > 0;
     }
   }
-  return v;
+  return a;
 }
 
 } // namespace sme