+++ sme/src/sme_simulationresult.cpp
@@ -17,7 +17,7 @@ void pybindSimulationResult(pybind11::module &m) {
                     float: the timepoint these simulation results are from
                     )")
       .def_readonly("concentration_image",
-                    &SimulationResult::concentrationImage,
+                    &SimulationResult::concentration_image,
                     R"(
                     list of list of list of int: an image of the species concentrations at this timepoint
 
@@ -25,7 +25,7 @@ void pybindSimulationResult(pybind11::module &m) {
                     concentration_image[y][x] = [r, g, b]
                     )")
       .def_readonly("species_concentration",
-                    &SimulationResult::speciesConcentration,
+                    &SimulationResult::species_concentration,
                     R"(
                     dict: the species concentrations at this timepoint
 
@@ -33,7 +33,7 @@ void pybindSimulationResult(pybind11::module &m) {
                     list of list of float, where species_concentration['A'][y][x]
                     is the concentration of species 'A" at the point (x,y)
                     )")
-      .def_readonly("species_dcdt", &SimulationResult::speciesDcdt,
+      .def_readonly("species_dcdt", &SimulationResult::species_dcdt,
                     R"(
                     dict: the rate of change of the species concentrations at this timepoint
 
@@ -53,7 +53,7 @@ std::string SimulationResult::getStr() const {
   std::string str("<sme.SimulationResult>\n");
   str.append(fmt::format("  - timepoint: {}\n", timePoint));
   str.append(
-      fmt::format("  - number of species: {}\n", speciesConcentration.size()));
+      fmt::format("  - number of species: {}\n", species_concentration.size()));
   return str;
 }